Enhanced tests Upated editor

+++ src/hooks/hooks.billing.js
@@ -1,5 +1,5 @@
 import makeDebug from 'debug'
-
+import _ from 'lodash'
 const debug = makeDebug('kalisio:kBilling:billing:hooks')
 
 export async function removeBilling (hook) {
@@ -7,18 +7,21 @@ export async function removeBilling (hook) {
     throw new Error(`The 'removeBilling' hook should only be used as a 'after' hook.`)
   }
 
-  let billingObjectId = hook.result._id
-  let billingObjectService = hook.service
-  let customerId = hook.result.billing.customer.id
-  debug('Removing billing from object ' + billingObjectId + ' of service ' + billingObjectService)
-  const billingService = hook.app.getService('billing')
-  await billingService.remove(customerId, {
-    query: {
-      action: 'customer',
-      billingObjectId: billingObjectId,
-      billingObjectService: billingObjectService
-    },
-    patch: !hook.result.deleted
-  })
+  let customer = hook.result.billing.customer
+  if (!_.isNil(customer)) {
+    let billingObjectId = hook.result._id
+    let billingObjectService = hook.service
+    let customerId = hook.result.billing.customer.id
+    debug('Removing billing from object ' + billingObjectId + ' of service ' + billingObjectService)
+    const billingService = hook.app.getService('billing')
+    await billingService.remove(customerId, {
+      query: {
+        action: 'customer',
+        billingObjectId: billingObjectId,
+        billingObjectService: billingObjectService
+      },
+      patch: false
+    })
+  }
   return hook
 }
